fix: address code review issues for Open File command
- Fix onModeChange callback not called when initialMode prop changes
- Fix stale closure in App.tsx by memoizing useNoteState return value
- Prevent infinite initialization loop in useNoteState with hasInitialized ref
- Consolidate duplicate mode-reset effects in CommandPalette
- Handle future timestamps in formatRelativeDate utility
- Extract magic number to MAX_RECENT_NOTES constant
- Add closeOnSelect property to Command type for explicit palette control
- Split CommandPalette.test.tsx into focused test files
- Replace setTimeout with requestAnimationFrame for input focus

Closes: scribe-x10, scribe-k2b, scribe-o5n, scribe-byj, scribe-7cv,
        scribe-a5x, scribe-ard, scribe-1i3, scribe-1if, scribe-8uf
